+++ src/test/java/org/springframework/data/couchbase/transactions/CouchbasePersonTransactionIntegrationTests.java
@@ -26,6 +26,7 @@ import static org.springframework.data.couchbase.transactions.util.TransactionTe
 
 import lombok.Data;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Disabled;
 import org.springframework.dao.OptimisticLockingFailureException;
 import org.springframework.data.couchbase.transactions.util.TransactionTestUtil;
 import reactor.core.publisher.Mono;
@@ -78,7 +79,7 @@ import com.couchbase.client.java.transactions.error.TransactionFailedException;
  * @author Michael Reiche
  */
 @IgnoreWhen(missesCapabilities = Capabilities.QUERY, clusterTypes = ClusterType.MOCKED)
-@SpringJUnitConfig(classes = { Config.class, PersonService.class })
+@SpringJUnitConfig(classes = { TransactionsConfigCouchbaseTransactionManager.class, PersonService.class })
 public class CouchbasePersonTransactionIntegrationTests extends JavaIntegrationTests {
 	// intellij flags "Could not autowire" when config classes are specified with classes={...}. But they are populated.
 	@Autowired CouchbaseClientFactory couchbaseClientFactory;
@@ -280,6 +281,7 @@ public class CouchbasePersonTransactionIntegrationTests extends JavaIntegrationT
 	/**
 	 * I think this test might fail sometimes?   Does it need retryWhen() ?
 	 */
+	@Disabled("todo gp: disabling temporarily as hanging intermittently")
 	@Test
 	public void wrapperReplaceWithCasConflictResolvedViaRetry() {
 		Person person = new Person(1, "Walter", "White");
